+++ src/components/signUp/SignUpBtn.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const SignUpBtn = () => {
+  return (
+    <button className="btn1 bg-_primary-50">
+      <span className="text-white flex justify-center items-center h-full">
+        가입하기
+      </span>
+    </button>
+  );
+};
+
+export default SignUpBtn;
